clean up chargemap folder, move functions/form files to corresponding folders, add jest.config.js file for testing
