+++ components/filtersForm.js
@@ -1,7 +1,7 @@
 import {Modal, Button, Form} from 'react-bootstrap';
 import React, {useState} from "react";
 
-export default function Filters2({filters, setFilters, onCloseClick}) {
+export default function FiltersForm({filters, setFilters, onCloseClick}) {
   const [show, setShow] = useState(false);
   const handleClose = () => setShow(false);
   const handleShow = () => setShow(true);
@@ -24,11 +24,12 @@ export default function Filters2({filters, setFilters, onCloseClick}) {
       for (var i = 0; i < checkboxesC.length; i++) {
         arrayC.push(checkboxesC[i].value);
       }
-    //  alert(arrayC);
+      // console.log("work now");
       setFilters({
         operators:arrayO,
         connections:arrayC
       });
+
       onCloseClick();
 
     //  console.log(arrayO);
