+++ helper_functions/jsonEscape.js
@@ -1,0 +1,3 @@
+export default function jsonEscape(str)  {
+  return str.replace(/\n/g, "\\\\n").replace(/\r/g, "\\\\r").replace(/\t/g, "\\\\t");
+}
