+++ package.json
@@ -7,12 +7,13 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "test": "jest"
+    "test": "jest --coverage"
   },
   "dependencies": {
     "@jest/globals": "^29.5.0",
     "@mapbox/mapbox-gl-geocoder": "^5.0.1",
     "@sendgrid/mail": "^7.7.0",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "api": "^5.0.7",
     "axios": "^1.3.4",
