+++ src/pages/ChargerMap/ChargerMap.js
@@ -1,7 +1,10 @@
 import Script from 'next/script';
 import {Helmet} from 'react-helmet';
-// import Filters from './filters.js';
-import Filters2 from './filters2.js';
+// import helper functions and form component
+import FiltersForm from '../../../components/filtersForm.js';
+import getOperatorsFilters from '../../../helper_functions/getOperatorsFilters.js'
+import getConnectionsFilters from '../../../helper_functions/getConnectionsFilters.js'
+import jsonEscape from '../../../helper_functions/jsonEscape.js'
 // import activity list
 import ActivityList from 'src/pages/ActivityList/ActivityList.js';
 
@@ -27,30 +30,8 @@ export default function ChargerMap(props) {
     connections:[]
   });
 
- function getOperatorsFilters (filterArray) {
-   let ofilters = [];
-   if(filterArray.operators.length!==0) {
-    ofilters=filters.operators.map((operator)=>{
-      return ['in', operator, ['string', ['get', 'poi']]];
-     });
-   }
-   return ofilters;
- }
-
- function getConnectionsFilters (filterArray) {
-  let cfilters = [];
-  if(filterArray.connections.length!==0) {
-    cfilters = filters.connections.map((connection)=>{
-      return ['in', connection, ['string', ['get', 'connectionType']]];
-  });
-
-  };
-
-  return cfilters;
-}
-
 function handleClick() {
-  console.log("in handle click");
+  // console.log("in handle click");
   let combinedFilters = getOperatorsFilters(filters).concat(getConnectionsFilters(filters));
   if (combinedFilters.length!== 0) {
     let filter = ['any',].concat(combinedFilters);
@@ -59,11 +40,6 @@ function handleClick() {
         // alert(JSON.stringify(filter));
   }
 }
-
-function jsonEscape(str)  {
-  return str.replace(/\n/g, "\\\\n").replace(/\r/g, "\\\\r").replace(/\t/g, "\\\\t");
-}
-
 
   useEffect(() => {
     if (map.current) return; // initialize map only once
@@ -192,7 +168,7 @@ function jsonEscape(str)  {
  Longitude: {lng} | Latitude: {lat} | Zoom: {zoom}
  </div> */}
   {/* <div> */}
-     <Filters2 filters={filters} setFilters={setFilters} onCloseClick={handleClick} />
+     <FiltersForm filters={filters} setFilters={setFilters} onCloseClick={handleClick} />
      <div ref={mapContainer} className="map-container" />
   </span>
   <pre id="quake-info">
